requirements.txt
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
cryptography==37.0.2
PyJWT==2.4.0
python-decouple==3.6